Fix warnings due to API changes in iOS 10, Beta 2
iOS 10, Beta 2 introduced changes to `CALayerDelegate` and
`CAAnimationDelegate`, both of which are now proper protocols instead of
informal ones.

This can be seen in the API diffs:
https://developer.apple.com/library/prerelease/content/releasenotes/General/iOS10APIDiffs/Objective-C/QuartzCore.html

This change fixes the warnings by introducing a macro which compiles to
those two delegate protocols when on the iOS 10 SDK but compiles to
placeholder protocols when using the iOS 9 SDK or below.